requirements.txt
# LangGraph City Events Agent
langgraph>=0.3.25
langchain>=0.3.14
langchain-core>=0.3.30
langchain-google-genai>=2.0.9
langchain-community>=0.3.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
tavily-python>=0.5.0
pyowm>=3.3.0
